fixed reroute

+++ views/ProductView.php
@@ -17,14 +17,16 @@ class ProductView
         echo "<a href='?page=editproduct&asignment=add'>Add clothes </a>";
         echo "<div class='row'>";
         foreach ($products as $current) {
-            $html = <<<HTML
-                <div class='col-4'>
-                    <p>$current[title]</p>
-                    <a href='?page=editproduct&asignment=edit&id=$current[product_id]'> edit </a>
-                    <a href='?page=editproduct&asignment=delete&id=$current[product_id]'> delete </a>
-                </div>
-                HTML;
-            echo $html;
+            if($current['available']){
+                $html = <<<HTML
+                    <div class='col-4'>
+                        <p>$current[title]</p>
+                        <a href='?page=editproduct&asignment=edit&id=$current[product_id]'> edit </a>
+                        <a href='?page=editproduct&asignment=delete&id=$current[product_id]'> delete </a>
+                    </div>
+                    HTML;
+                echo $html;
+            }
         }
         echo "</div>";
     }
